requirements.txt
pillow
ImageHash